+++ packages/tables/src/Table.php
@@ -41,7 +41,6 @@ class Table extends ViewComponent
     use Table\Concerns\HasRecords;
     use Table\Concerns\HasRecordUrl;
     use Table\Concerns\HasToolbarActions;
-    use Table\Concerns\CanScrollToTop;
 
     /**
      * @var view-string
